Add two more test cases for string compress coding challenge.
I believe that the existing test cases are not strict enough to assure
a proper solution for this challenge.

I was able to implement a wrong, but still passing solution to this
basing on set and collections.Counter mechanism, since the test cases
present only situations with singular occurrences of a letter chain and
all the test strings are alphabetically sorted.

I believe that the two new test cases enforce more thorough approach to
the problem and, in effect, the test cases are more descriptive.