Include `Counter`-based permutation solution
Summary:
The `collections` module exposes a type `Counter`, which makes the code
much more readable for this solution. It would be a one-liner if not for
the `None`-checking.

Test Plan:
Existing unit tests suffice for the new implementation.

The `test_permutation_solution.py` file still fails when run by itself,
for the same reason as before: not only are alternate solutions only in
the notebook, but the `Permutations` solution is as well. Thus, there is
a `NameError` on line 18. If you comment this line out, it still "works"
(meaning, it still silently absorbs the subsequent `NameError`s).

Please note that I tested this under Python 2 within Jupyter.
I also pasted the function into a Python 3 REPL and unit-tested it
manually, but I don't have Jupyter for Python 3.