Refactor code
Inputs don't need to be sorted, but we have to change how the bottom up implementation populates the results array.

We can also simplify some of the statements by:

1) removing the call to list when we create the list 'items', as there's no need for it;
2) initializing the list items with zeros; and
3) removing the if statement in the nested loop, since it's not necessary anymore because of #2.

We have to change the test too. We can't hardcode the labels to indices because they may be in a different position.